US market dataset to be used in the Neural Network
This dataset considers the N=43 assets we decided to use for the US market. 
Columns are windows, rows are data indicated below:

The first N**2 rows are the Euclidean matrix.
The next N**2 rows are the Gaussian matrix.
The next N**2 rows are the Laplacian matrix.
The next N rows are the cluster labels.
The last 3 rows are the one-hot vector which indicate the optimal strategy for the given window (obtained using Theo's results).